requirements.txt
# Core framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LlamaIndex
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-llms-openai>=0.1.0
llama-index-vector-stores-qdrant>=0.3.0

# LlamaParse (optional – set LLAMA_CLOUD_API_KEY to enable)
llama-parse>=0.4.0

# Vector store
qdrant-client>=1.9.0

# Embeddings
sentence-transformers>=3.0.0
FlagEmbedding>=1.2.0
transformers>=4.40.0,<5.0.0

# HTML parsing fallback
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0

# OpenAI
openai>=1.30.0

